requirements.txt
click==6.6
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pyaml==16.9.0
requests==2.18.4
six==1.10.0
slackclient==1.0.5
websocket-client==0.37.0
Werkzeug==0.11.11
jira==1.0.10
oauthlib==1.0.3
PyJWT==1.3.0
PyYAML==3.11
requests-oauthlib==0.8.0
urllib3==1.22
requests-toolbelt==0.8.0
cryptography==1.2.3
